requirements.txt
fastapi
uvicorn
python-multipart
llama-cpp-python
python-docx
pdfplumber